WIP: Update many tests
