+++ property_tests/tests/int/test/generate_tests.ts
@@ -1,0 +1,35 @@
+import { deepEqual } from 'fast-equals';
+
+import { getActor } from 'azle/property_tests';
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { Test } from 'azle/test';
+
+export function generateTests(
+    functionName: string,
+    namedParamInts: Named<CandidMeta<bigint>>[],
+    returnInt: CandidMeta<bigint>
+): Test[] {
+    const expectedResult = namedParamInts.reduce(
+        (acc, param) => acc + param.el.agentResponseValue,
+        returnInt.agentResponseValue
+    );
+    const paramValues = namedParamInts.map(
+        (param) => param.el.agentArgumentValue
+    );
+
+    return [
+        {
+            name: `int ${functionName}`,
+            test: async () => {
+                const actor = getActor('./tests/int/test');
+
+                const result = await actor[functionName](...paramValues);
+
+                return {
+                    Ok: deepEqual(result, expectedResult)
+                };
+            }
+        }
+    ];
+}
