+++ property_tests/tests/int16/test/generate_body.ts
@@ -1,0 +1,30 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamInt16s: Named<CandidMeta<number>>[],
+    returnInt16: CandidMeta<number>
+): string {
+    const paramsAreNumbers = namedParamInt16s
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'number') throw new Error('${param.name} must be a number');`;
+        })
+        .join('\n');
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamInt16s);
+
+    const sum = namedParamInt16s.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnInt16.src.valueLiteral);
+    const count = namedParamInt16s.length + 1;
+    const average = `Math.floor((${sum}) / ${count})`;
+
+    return `
+        ${paramsAreNumbers}
+
+        ${paramsCorrectlyOrdered}
+
+        return ${average};
+    `;
+}
