+++ property_tests/tests/nat/test/generate_body.ts
@@ -1,0 +1,28 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamNats: Named<CandidMeta<bigint>>[],
+    returnNat: CandidMeta<bigint>
+): string {
+    const paramsAreBigInts = namedParamNats
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'bigint') throw new Error('${param.name} must be a bigint');`;
+        })
+        .join('\n');
+
+    const sum = namedParamNats.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnNat.src.valueLiteral);
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamNats);
+
+    return `
+        ${paramsCorrectlyOrdered}
+
+        ${paramsAreBigInts}
+
+        return ${sum};
+    `;
+}
