+++ property_tests/tests/nat16/test/generate_body.ts
@@ -1,0 +1,30 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamNat16s: Named<CandidMeta<number>>[],
+    returnNat16: CandidMeta<number>
+): string {
+    const paramsAreNumbers = namedParamNat16s
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'number') throw new Error('${param.name} must be a number');`;
+        })
+        .join('\n');
+
+    const sum = namedParamNat16s.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnNat16.src.valueLiteral);
+    const count = namedParamNat16s.length + 1;
+    const average = `Math.floor((${sum}) / ${count})`;
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamNat16s);
+
+    return `
+        ${paramsAreNumbers}
+
+        ${paramsCorrectlyOrdered}
+
+        return ${average};
+    `;
+}
