+++ property_tests/tests/nat32/test/generate_body.ts
@@ -1,0 +1,30 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamNat32s: Named<CandidMeta<number>>[],
+    returnNat32: CandidMeta<number>
+): string {
+    const paramsAreNumbers = namedParamNat32s
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'number') throw new Error('${param.name} must be a number');`;
+        })
+        .join('\n');
+
+    const sum = namedParamNat32s.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnNat32.src.valueLiteral);
+    const count = namedParamNat32s.length + 1;
+    const average = `Math.floor((${sum}) / ${count})`;
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamNat32s);
+
+    return `
+        ${paramsAreNumbers}
+
+        ${paramsCorrectlyOrdered}
+
+        return ${average};
+    `;
+}
