+++ property_tests/tests/nat64/test/generate_body.ts
@@ -1,0 +1,30 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamNat64s: Named<CandidMeta<bigint>>[],
+    returnNat64: CandidMeta<bigint>
+): string {
+    const paramsAreBigInts = namedParamNat64s
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'bigint') throw new Error('${param.name} must be a bigint');`;
+        })
+        .join('\n');
+
+    const sum = namedParamNat64s.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnNat64.src.valueLiteral);
+    const count = namedParamNat64s.length + 1;
+    const average = `(${sum}) / ${count}n`;
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamNat64s);
+
+    return `
+        ${paramsAreBigInts}
+
+        ${paramsCorrectlyOrdered}
+
+        return ${average};
+    `;
+}
