+++ property_tests/tests/nat8/test/generate_body.ts
@@ -1,0 +1,30 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamNat8s: Named<CandidMeta<number>>[],
+    returnNat8: CandidMeta<number>
+): string {
+    const paramsAreNumbers = namedParamNat8s
+        .map((param) => {
+            return `if (typeof ${param.name} !== 'number') throw new Error('${param.name} must be a number');`;
+        })
+        .join('\n');
+
+    const sum = namedParamNat8s.reduce((acc, { name }) => {
+        return `${acc} + ${name}`;
+    }, returnNat8.src.valueLiteral);
+    const count = namedParamNat8s.length + 1;
+    const average = `Math.floor((${sum}) / ${count})`;
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamNat8s);
+
+    return `
+        ${paramsAreNumbers}
+
+        ${paramsCorrectlyOrdered}
+
+        return ${average};
+    `;
+}
