+++ property_tests/tests/null/test/test.ts
@@ -1,12 +1,12 @@
 import fc from 'fast-check';
-import { deepEqual } from 'fast-equals';
 
-import { CanisterArb } from '../../../arbitraries/canister_arb';
-import { NullArb } from '../../../arbitraries/candid/primitive/null';
-import { getActor, runPropTests } from '../../../../property_tests';
-import { CandidMeta } from '../../../arbitraries/candid/candid_arb';
-import { Test } from '../../../../test';
-import { Named, QueryMethodArb } from '../../../arbitraries/query_method_arb';
+import { runPropTests } from 'azle/property_tests';
+import { NullArb } from 'azle/property_tests/arbitraries/candid/primitive/null';
+import { CanisterArb } from 'azle/property_tests/arbitraries/canister_arb';
+import { QueryMethodArb } from 'azle/property_tests/arbitraries/query_method_arb';
+
+import { generateBody } from './generate_body';
+import { generateTests } from './generate_tests';
 
 const AllNullsQueryMethod = QueryMethodArb(fc.array(NullArb), NullArb, {
     generateBody,
@@ -14,45 +14,3 @@ const AllNullsQueryMethod = QueryMethodArb(fc.array(NullArb), NullArb, {
 });
 
 runPropTests(CanisterArb(AllNullsQueryMethod));
-
-function generateBody(
-    namedParamNulls: Named<CandidMeta<null>>[],
-    returnNull: CandidMeta<null>
-): string {
-    const areAllNull = namedParamNulls.reduce((acc, { name }) => {
-        return `${acc} && ${name} === null`;
-    }, 'true');
-
-    const allNullCheck = `if (!(${areAllNull})) throw new Error("Not all of the values were null")`;
-
-    return `
-        ${allNullCheck}
-
-        return ${returnNull.src.valueLiteral};
-    `;
-}
-
-function generateTests(
-    functionName: string,
-    namedParamNulls: Named<CandidMeta<null>>[],
-    _returnNull: CandidMeta<null>
-): Test[] {
-    return [
-        {
-            name: `test ${functionName}`,
-            test: async () => {
-                const actor = getActor('./tests/null/test');
-
-                const result = await actor[functionName](
-                    ...namedParamNulls.map(
-                        (param) => param.el.agentArgumentValue
-                    )
-                );
-
-                return {
-                    Ok: deepEqual(result, null)
-                };
-            }
-        }
-    ];
-}
