+++ property_tests/tests/opt/test/generate_body.ts
@@ -1,0 +1,32 @@
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Opt } from 'azle/property_tests/arbitraries/candid/constructed/opt_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+
+export function generateBody(
+    namedParamOpts: Named<CandidMeta<Opt>>[],
+    returnOpt: CandidMeta<Opt>
+): string {
+    const areParamsOpts = namedParamOpts
+        .map((param) => {
+            return `if (!${isParamOpt(param.name)}) throw new Error('${
+                param.name
+            } must be an Opt');`;
+        })
+        .join('\n');
+
+    const paramsCorrectlyOrdered = areParamsCorrectlyOrdered(namedParamOpts);
+
+    const returnStatement = returnOpt.src.valueLiteral;
+
+    return `
+        ${areParamsOpts}
+        ${paramsCorrectlyOrdered}
+
+        return ${returnStatement};
+    `;
+}
+
+function isParamOpt(paramName: string): string {
+    return `(${paramName}.Some !== undefined || ${paramName}.None !== undefined)`;
+}
