+++ property_tests/tests/opt/test/generate_tests.ts
@@ -1,0 +1,34 @@
+import { deepEqual } from 'fast-equals';
+
+import { getActor } from 'azle/property_tests';
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
+import { Opt } from 'azle/property_tests/arbitraries/candid/constructed/opt_arb';
+import { Named } from 'azle/property_tests/arbitraries/query_method_arb';
+import { Test } from 'azle/test';
+
+export function generateTests(
+    functionName: string,
+    namedParamOpts: Named<CandidMeta<Opt>>[],
+    returnOpt: CandidMeta<Opt>
+): Test[] {
+    const expectedResult = returnOpt.agentResponseValue;
+
+    return [
+        {
+            name: `opt ${functionName}`,
+            test: async () => {
+                const actor = getActor('./tests/opt/test');
+
+                const params = namedParamOpts.map(
+                    (param) => param.el.agentArgumentValue
+                );
+
+                const result = await actor[functionName](...params);
+
+                return {
+                    Ok: deepEqual(expectedResult, result)
+                };
+            }
+        }
+    ];
+}
