+++ property_tests/tests/variant/test/test.ts
@@ -1,16 +1,16 @@
 import fc from 'fast-check';
 import { deepEqual } from 'fast-equals';
 
-import { areParamsCorrectlyOrdered } from '../../../are_params_correctly_ordered';
-import { CandidMeta } from '../../../arbitraries/candid/candid_arb';
+import { areParamsCorrectlyOrdered } from 'azle/property_tests/are_params_correctly_ordered';
+import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
 import {
     Variant,
     VariantArb
-} from '../../../arbitraries/candid/constructed/variant_arb';
-import { CanisterArb } from '../../../arbitraries/canister_arb';
-import { QueryMethodArb } from '../../../arbitraries/query_method_arb';
-import { getActor, runPropTests } from '../../../../property_tests';
-import { Test } from '../../../../test';
+} from 'azle/property_tests/arbitraries/candid/constructed/variant_arb';
+import { CanisterArb } from 'azle/property_tests/arbitraries/canister_arb';
+import { QueryMethodArb } from 'azle/property_tests/arbitraries/query_method_arb';
+import { getActor, runPropTests } from 'azle/property_tests';
+import { Test } from 'azle/test';
 
 const UniqueVariantsArray = fc.uniqueArray(VariantArb, {
     selector: (entry) => entry.src.candidType
